Fix invalid type comparison.

+++ uint64.go
@@ -15,7 +15,7 @@ func Uint64(f *frm.Field, inp ...string) {
 		return
 	}
 	value := f.Uint64()
-	if f.Min != nil && value < f.Min.(uint) || f.Max != nil && value > f.Max.(uint) {
+	if f.Min != nil && value < f.Min.(uint64) || f.Max != nil && value > f.Max.(uint64) {
 		f.Err = fmt.Sprintf("Must be between %v and %v.", f.Min, f.Max)
 		return
 	}
